refactor(Day6): move to single file with Container class
